Moved Game to it's own file and wrote tests for the app
I moved Game and getWinningSquares to their own file and removed them form index.js. I then wrote tests to 1.Check if the Squares are rendering properly 2. Check if the board is rendering properly 3. Check if the Square class name changed when it is labeled a winner 4. Check if the getWinningSquares method is returning the correct values for a winning situation, a draw, or when the game is still in progress