Added bootstrap components
The sort button was changed to a bootstrap button, moved to the left side of the screen, and changed color. The X's and O's as well as the status and moves are now white with the X's and O's now changing to black when there is a winner. A picture of the desert was added as the background image of the page and two glyphicons were added to the left side.